python/requirements-build.txt
pyinstaller>=6.0,<7.0
pillow>=10.0,<12.0
requests>=2.28,<3.0